created all pages and added checks for user login status to redirect if logged out
